CADASTRANDO E LISTANDO, SÓ FALTA EDITAR E EXCLUIR

+++ Presentation.Biblioteca/Models/LivroViewModel.cs
@@ -12,7 +12,7 @@ namespace MVC.Biblioteca.Models
         public string ISBN { get; set; }
         public int idAutor { get; set; }
         public DateTime dataPubLivro { get; set; }
-        public double precoLivro { get; set; }
+        public decimal precoLivro { get; set; }
         public int idEditora { get; set; }
 
         public string DataFormatada => dataPubLivro.ToShortDateString();
